Moved overlay source to a createSource function
