+++ package/src/routes/overlay/editor/component/Source.svelte
@@ -1,12 +1,26 @@
+<script lang="ts" context="module">
+  import { field } from "$lib/overlay/index.js";
+
+  export const label = "testing";
+  export const transform = {
+    x: 100,
+    y: 100,
+    width: 200,
+    height: 200,
+  };
+
+  export const options = {
+    color: field("text", "red"),
+    text: field("multiline", "Hello\nWorld!"),
+  };
+</script>
+
 <script lang="ts">
   export let color: string;
   export let text: string;
 </script>
 
-<div
-  class="source"
-  style:background-color={color}
->
+<div class="source" style:background-color={color}>
   <pre>{text}</pre>
 </div>
